moved registration route to a new file
